Call cleanup of insertion effects when hidden
Insertion effects do not unmount when a subtree goes offscreen, this means we still have to traverse the subtree in that case.

We could potentially use a new subtree bit flag to avoid the traversal if there's no insertion effects, but I'm not sure that's worth the tradeoff of using a new flag?

Likely also fixes https://github.com/facebook/react/issues/26670